+++ pyairtable/models/__init__.py
@@ -1,0 +1,5 @@
+from .comment import Comment
+
+__all__ = [
+    "Comment",
+]